requirements.txt
streamlit==1.26.1
faiss-cpu==1.7.4
transformers==4.36.2
sentence-transformers==2.2.2
PyPDF2==3.0.1
langchain==0.1.0
python-dotenv==1.0.1